+++ src/XGD.h
@@ -13,13 +13,13 @@ namespace XGD {
 
     constexpr char     NAME[]         = XGDTOOL_NAME;
     constexpr char     VERSION[]      = XGDTOOL_VERSION;
-    constexpr uint32_t VERSION_LEN    = sizeof(VERSION) - 1;
-    constexpr uint32_t BUFFER_SIZE    = 0x10000; // 64KB
+    constexpr uint64_t VERSION_LEN    = sizeof(VERSION) - 1;
+    constexpr uint64_t BUFFER_SIZE    = 0x10000; // 64KB
 
     constexpr char     OPTIMIZED_TAG[]       = "in!xgdt!" XGDTOOL_VERSION;
-    constexpr uint32_t OPTIMIZED_TAG_OFFSET  = 31337;
-    constexpr uint32_t OPTIMIZED_TAG_LEN     = 8 + VERSION_LEN;
-    constexpr uint32_t OPTIMIZED_TAG_LEN_MIN = 7;
+    constexpr uint64_t OPTIMIZED_TAG_OFFSET  = 31337;
+    constexpr uint64_t OPTIMIZED_TAG_LEN     = 8 + VERSION_LEN;
+    constexpr uint64_t OPTIMIZED_TAG_LEN_MIN = 7;
 
 };
